Tensorflow
Tensorflow is an open source library for numerical computations and large scale Machine Learning.

TENSOR is a vector or a matrix of n-dimension. Flow means the flow of operations.

It is used mainly in Object Detection,Fraud Detection.

Steps in Machine learning are:

1. Training data
2. Model
3. Cost Function
4. Optimization Techniques
5. Evaluation Criteria
Estimator is a framework to build a new Machine Learning Model.